Express middleware now can verify passwords from frontend against RDS
I still need to implement the case of the user not existing. The frontend needs a redirect in the case of a successful login and suitable errors in the case of unsuccessful attempts. The dummy accounts that I have made are: DjAlex-password1, CalamityClark-password1, Tmoney-password1, BigBodine-pword. I will give the super secret json file with db credentials upon request so you cif anyone wants to play around with it